// apps/admin/web/src/types/garage.ts
export interface BucketKeyPerm {
  accessKeyId: string;
  name: string;
  permissions: {
    read: boolean;
    write: boolean;
    owner: boolean;
  };
  bucketLocalAliases?: string[];
}